+++ internal/sdkprovider/provider/provider.go
@@ -181,9 +181,11 @@ func Provider(version string) *schema.Provider {
 			"aiven_account_authentication": account.ResourceAccountAuthentication(),
 
 			// organization
-			"aiven_organizational_unit":     organization.ResourceOrganizationalUnit(),
-			"aiven_organization_user":       organization.ResourceOrganizationUser(),
-			"aiven_organization_user_group": organization.ResourceOrganizationUserGroup(),
+			"aiven_organization":                   organization.ResourceOrganization(),
+			"aiven_organizational_unit":            organization.ResourceOrganizationalUnit(),
+			"aiven_organization_user":              organization.ResourceOrganizationUser(),
+			"aiven_organization_user_group":        organization.ResourceOrganizationUserGroup(),
+			"aiven_organization_user_group_member": organization.ResourceOrganizationUserGroupMember(),
 
 			// project
 			"aiven_project":       project.ResourceProject(),