+++ internal/sdkprovider/service/organization/organization_user_group_data_source.go
@@ -19,7 +19,7 @@ import (
 func DatasourceOrganizationUserGroup() *schema.Resource {
 	return &schema.Resource{
 		ReadContext: datasourceOrganizationUserGroupRead,
-		Description: "The Organization User Groupe data source provides information about the existing Aiven" +
+		Description: "The Organization User Group data source provides information about the existing Aiven" +
 			" Organization User Group.",
 		Schema: schemautil.ResourceSchemaAsDatasourceSchema(
 			aivenOrganizationUserGroupSchema, "organization_id", "name",